backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
cachetools==7.0.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
chess==1.11.2
click==8.3.1
cryptography==46.0.5
cuda-bindings==12.9.4
cuda-pathfinder==1.4.0
fastapi==0.135.1
filelock==3.25.0
fsspec==2026.2.0
google-api-core==2.30.0
google-api-python-client==2.191.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
googleapis-common-protos==1.72.0
grpcio==1.78.0
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httplib2==0.31.2
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.5.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
langgraph
langchain
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx
numpy
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
oauth2client==4.1.3
opencv-python==4.13.0.92
orjson==3.11.7
packaging==26.0
pandas
pillow
proto-plus==1.27.1
protobuf
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic
pydantic_core==2.41.5
Pygments==2.19.2
pymilvus==2.6.9
pyparsing==3.3.2
python-chess==1.999
python-dateutil==2.9.0.post0
python-dotenv
python-multipart
PyYAML==6.0.3
regex==2026.2.28
requests==2.32.5
rich==14.3.3
rsa==4.9.1
redis
safetensors==0.7.0
scikit-learn
scipy
sentence-transformers
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
starlette==0.52.1
stockfish==4.0.8
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch
tqdm==4.67.3
transformers
triton==3.6.0
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
